Conditionally support `no_std` and `std` for macro imports
This commit uses conditionals to support `no_std` and `std` for the
macro impl crate.

When the `std` feature is missing, we import `core` and `alloc`,
and use their implementations of `std::mem`, `std::os::raw::c_void`, and
`std::string::String`.

We also use `core2` when `std` is missing, and this allows for a `no_std`
compatible implementation of `std::io`.

When, and this is by default, the `std` feature
is present, we just use the standard `std` implementations of the
aforementioned features.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>